refactoring add middleware validation for function Add
