+++ models/contacts.json
@@ -52,5 +52,11 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
+  },
+  {
+    "name": "All Ray",
+    "email": "nulla.ante@vestibul.co.ff",
+    "phone": "(992) 111-1111",
+    "id": "23c9ab03-ecbd-469c-a93a-46358e27f2de"
   }
 ]
