+++ schemas/index.js
@@ -1,0 +1,5 @@
+const contactSchema = require("./contact");
+
+module.exports = {
+  contactSchema,
+};
